Feature/7

+++ src/routes/AdminRoutes.ts
@@ -3,19 +3,25 @@ import { IReq, IRes } from './common/types'
 import HttpStatusCodes from '@src/common/HttpStatusCodes'
 import check from './common/check'
 import Admin from '@src/models/Admin'
+import Authorization from '@src/util/Authorization'
+import { UserType } from '@src/models/User'
 
-const getAll = async (_: IReq, res: IRes) => {
+const getAll = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAuthorized(req, res, [UserType.ADMIN])) return
   const admins = await AdminService.getAll()
-  res.status(HttpStatusCodes.OK).json({ admins })
+  const publicAdmins = admins.map(admin => Admin.ObtainPublicAdmin(admin))
+  res.status(HttpStatusCodes.OK).json({ admins: publicAdmins })
 }
 
 const add = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAuthorized(req, res, [UserType.ADMIN])) return
   const admin = check.isValid(req.body, 'admin', Admin.isAdmin)
   await AdminService.addOne(admin)
   res.status(HttpStatusCodes.CREATED).end()
 }
 
 const update = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAuthorized(req, res, [UserType.ADMIN])) return
   const id = Number(req.params.id)
   const admin = check.isValid(req.body, 'admin', Admin.isPartialAdmin)
   await AdminService.updateOne(id, admin)
@@ -23,6 +29,7 @@ const update = async (req: IReq, res: IRes) => {
 }
 
 const delete_ = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAuthorized(req, res, [UserType.ADMIN])) return
   const id = Number(req.params.id)
   await AdminService.delete(id)
   res.status(HttpStatusCodes.OK).end()
