+++ src/routes/CustomerRoutes.ts
@@ -3,19 +3,25 @@ import { IReq, IRes } from './common/types'
 import HttpStatusCodes from '@src/common/HttpStatusCodes'
 import check from './common/check'
 import Customer from '@src/models/Customer'
+import Authorization from '@src/util/Authorization'
+import { UserType } from '@src/models/User'
 
-const getAll = async (_: IReq, res: IRes) => {
+const getAll = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAuthorized(req, res, [UserType.ADMIN])) return
   const customers = await CustomerService.getAll()
-  res.status(HttpStatusCodes.OK).json({ customers })
+  const publicCustomers = customers.map(customer => Customer.ObtainPublicCustomer(customer))
+  res.status(HttpStatusCodes.OK).json({ customers: publicCustomers })
 }
 
 const add = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAuthorized(req, res, [UserType.ADMIN])) return
   const customer = check.isValid(req.body, 'customer', Customer.isCustomer)
   await CustomerService.addOne(customer)
   res.status(HttpStatusCodes.CREATED).end()
 }
 
 const update = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAuthorized(req, res, [UserType.ADMIN])) return
   const id = Number(req.params.id)
   const customer = check.isValid(req.body, 'customer', Customer.isPartialCustomer)
   await CustomerService.updateOne(id, customer)
@@ -23,6 +29,7 @@ const update = async (req: IReq, res: IRes) => {
 }
 
 const delete_ = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAuthorized(req, res, [UserType.ADMIN])) return
   const id = Number(req.params.id)
   await CustomerService.delete(id)
   res.status(HttpStatusCodes.OK).end()
