+++ src/routes/ReservationRoutes.ts
@@ -5,19 +5,24 @@ import check from './common/check'
 import Reservation from '@src/models/Reservation'
 import Vehicle from '@src/models/Vehicle'
 import { dateValidaton } from '@src/util'
+import Authorization from '@src/util/Authorization'
+import { UserType } from '@src/models/User'
 
-const getAll = async (_: IReq, res: IRes) => {
+const getAll = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAuthorized(req, res, [UserType.ADMIN, UserType.EMPLOYEE])) return
   const reservations = await ReservationService.getAll()
   res.status(HttpStatusCodes.OK).json({ reservations })
 }
 
 const add = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAuthorized(req, res, [UserType.ADMIN])) return
   const reservation = check.isValid(req.body, 'reservation', Reservation.isReservation)
   await ReservationService.addOne(reservation)
   res.status(HttpStatusCodes.CREATED).end()
 }
 
 const update = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAuthorized(req, res, [UserType.ADMIN])) return
   const id = Number(req.params.id)
   const reservation = check.isValid(req.body, 'reservation', Reservation.isPartialReservation)
   await ReservationService.updateOne(id, reservation)
@@ -25,12 +30,14 @@ const update = async (req: IReq, res: IRes) => {
 }
 
 const delete_ = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAuthorized(req, res, [UserType.ADMIN])) return
   const id = Number(req.params.id)
   await ReservationService.delete(id)
   res.status(HttpStatusCodes.OK).end()
 }
 
 const reserve = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAuthorized(req, res, [UserType.ADMIN, UserType.EMPLOYEE, UserType.CUSTOMER])) return
   const vehicle = check.isValid(req.body, 'vehicle', Vehicle.isVehicle)
   const startTime = check.isValid(req.body, 'startTime', dateValidaton)
   const endTime = check.isValid(req.body, 'endTime', dateValidaton)
